omap: serial: disable serial port irq during suspend
If the serial port interrupt is left enabled during suspend, it
can fire during uart_resume_port, while the uart is in an
undefined state and the port lock is not held.

Disable the interrupt in serial_omap_suspend, and enable it in
serial_omap_resume.  This also requires moving the request_irq
into probe, instead of startup, so that the irq is always
available to disable.

This will save additional power in idle mode/deep sleep. It has beend tested and after this patch we had a better idle line e.g. in Battery Widget Monitor.

Change-Id: I7816d932f6991dd2d2e9eae021c627e058dc4aaa