Suspend Governor: Added symsearch (unexported symbol finder) to bind an exisiting function into suspend_gov module - this prevents any conflicts with sgx-sources - or linker errors in kernel source.
Before this we could solve the problem from kernel side - the related function was now declared as non-static so we could export the symbol or at least bind them via cpufreq.h. But the problem here was that when it came to compiling of the sgx sources, which are external built with Android Rom (but still dependant to the kernel source of course), it was conflicting with a similar funcntion declared as static in the sgx source. A declaration wit non-static didn't help.

Symsearch is our solution here.