Add Kernel Builder Script [WIP]
This is a big one. It will manage everything as much automated as possible. I plan to make this a little program also for noobs to being able to compile the kernel easily. The script will do almost everything....when it's done! For now it only supports Spyder. Other devices will follow ver soon